correções de autorização + refazendo Layout

+++ src/Controller/SchedulesController.php
@@ -53,12 +53,18 @@ class SchedulesController extends AppController
      */
     public function view($id = null)
     {
-        $schedule = $this->Schedules->get($id, [
-            'contain' => ['Users', 'Services']
-        ]);
+        if(is_null($id)){
+            
+            return $this->redirect(['action' => 'index']);
+        
+        }else{
+            $schedule = $this->Schedules->get($id, [
+                'contain' => ['Users', 'Services']
+            ]);
 
-        $this->set('schedule', $schedule);
-        $this->set('_serialize', ['schedule']);
+            $this->set('schedule', $schedule);
+            $this->set('_serialize', ['schedule']);
+        }
     }
 
     /**
@@ -93,22 +99,25 @@ class SchedulesController extends AppController
      */
     public function edit($id = null)
     {
-        $schedule = $this->Schedules->get($id, [
-            'contain' => []
-        ]);
-        if ($this->request->is(['patch', 'post', 'put'])) {
-            $schedule = $this->Schedules->patchEntity($schedule, $this->request->getData());
-            if ($this->Schedules->save($schedule)) {
-                $this->Flash->success(__('The schedule has been saved.'));
-
-                return $this->redirect(['action' => 'index']);
-            }
-            $this->Flash->error(__('The schedule could not be saved. Please, try again.'));
+        if(is_null($id)){
+            return $this->redirect(['action' => 'index']);
         }
-        $users = $this->Schedules->Users->find('list', ['limit' => 200]);
-        $services = $this->Schedules->Services->find('list', ['limit' => 200]);
-        $this->set(compact('schedule', 'users', 'services'));
-        $this->set('_serialize', ['schedule']);
+        else{
+            $schedule = $this->Schedules->get($id, ['contain' => []]);
+            if ($this->request->is(['patch', 'post', 'put'])) {
+                $schedule = $this->Schedules->patchEntity($schedule, $this->request->getData());
+                if ($this->Schedules->save($schedule)) {
+                    $this->Flash->success(__('The schedule has been saved.'));
+
+                    return $this->redirect(['action' => 'index']);
+                }
+                $this->Flash->error(__('The schedule could not be saved. Please, try again.'));
+            }
+            $users = $this->Schedules->Users->find('list', ['limit' => 200]);
+            $services = $this->Schedules->Services->find('list', ['limit' => 200]);
+            $this->set(compact('schedule', 'users', 'services'));
+            $this->set('_serialize', ['schedule']);
+        }
     }
 
     /**
@@ -120,14 +129,21 @@ class SchedulesController extends AppController
      */
     public function delete($id = null)
     {
-        $this->request->allowMethod(['post', 'delete']);
-        $schedule = $this->Schedules->get($id);
-        if ($this->Schedules->delete($schedule)) {
-            $this->Flash->success(__('The schedule has been deleted.'));
-        } else {
-            $this->Flash->error(__('The schedule could not be deleted. Please, try again.'));
-        }
+        if(is_null($id)){
+            return $this->redirect(['action' => 'index']);
+        }else{
+            if($this->request->is('get')){
+                return $this->redirect(['action' => 'index']);
+            }
+            $this->request->allowMethod(['post', 'delete']);
+            $schedule = $this->Schedules->get($id);
+            if ($this->Schedules->delete($schedule)) {
+                $this->Flash->success(__('The schedule has been deleted.'));
+            } else {
+                $this->Flash->error(__('The schedule could not be deleted. Please, try again.'));
+            }
 
-        return $this->redirect(['action' => 'index']);
+            return $this->redirect(['action' => 'index']);
+        }
     }
 }
