+++ src/Controller/UsersController.php
@@ -15,7 +15,16 @@ use Cake\Event\Event;
 class UsersController extends AppController
 {
 
+    public function initialize(){
+        parent::initialize();
+    }
 
+    public function beforeFilter(Event $event)
+    {
+        parent::beforeFilter($event);
+        $this->Auth->allow(['add', 'logout']);
+        $this->Auth->deny(['edit', 'index','view','delete']);
+    }
 
     public function login()
     {
@@ -34,12 +43,6 @@ class UsersController extends AppController
 
     public function logout(){
         return $this->redirect($this->Auth->logout());
-    }
-
-    public function beforeFilter(Event $event)
-    {
-        parent::beforeFilter($event);
-        $this->Auth->allow(['add', 'logout']);
     }
 
     /**
@@ -64,12 +67,20 @@ class UsersController extends AppController
      */
     public function view($id = null)
     {
-        $user = $this->Users->get($id, [
-            'contain' => ['Schedules']
-        ]);
 
-        $this->set('user', $user);
-        $this->set('_serialize', ['user']);
+        if(is_null($id)){
+            
+            return $this->redirect(['action' => 'index']);
+        
+        }else{
+
+            $user = $this->Users->get($id, [
+                'contain' => ['Schedules']
+            ]);
+
+            $this->set('user', $user);
+            $this->set('_serialize', ['user']);
+        }
     }
 
     /**
@@ -102,20 +113,28 @@ class UsersController extends AppController
      */
     public function edit($id = null)
     {
-        $user = $this->Users->get($id, [
-            'contain' => []
-        ]);
-        if ($this->request->is(['patch', 'post', 'put'])) {
-            $user = $this->Users->patchEntity($user, $this->request->getData());
-            if ($this->Users->save($user)) {
-                $this->Flash->success(__('Usuário editado com sucesso.'));
 
-                return $this->redirect(['action' => 'index']);
+        if(is_null($id)){
+            
+            return $this->redirect(['action' => 'index']);
+        
+        }else{
+
+            $user = $this->Users->get($id, [
+                'contain' => []
+            ]);
+            if ($this->request->is(['patch', 'post', 'put'])) {
+                $user = $this->Users->patchEntity($user, $this->request->getData());
+                if ($this->Users->save($user)) {
+                    $this->Flash->success(__('Usuário editado com sucesso.'));
+
+                    return $this->redirect(['action' => 'index']);
+                }
+                $this->Flash->error(__('O usuário não pode ser editado. Por favor, tente novamente.'));
             }
-            $this->Flash->error(__('O usuário não pode ser editado. Por favor, tente novamente.'));
+            $this->set(compact('user'));
+            $this->set('_serialize', ['user']);
         }
-        $this->set(compact('user'));
-        $this->set('_serialize', ['user']);
     }
 
     /**
@@ -127,16 +146,25 @@ class UsersController extends AppController
      */
     public function delete($id = null)
     {
-        $this->request->allowMethod(['post', 'delete']);
-        $user = $this->Users->get($id);
-        if ($this->Users->delete($user)) {
-            $this->Flash->success(__('Usuário deletado com sucesso.'));
-        } else {
-            $this->Flash->error(__('O usuário não pode ser deletado. Por favor, tente novamente.'));
+        if(is_null($id)){
+        
+            return $this->redirect(['action' => 'index']);
+        
+        }else{
+
+            if($this->request->is('get')){
+                return $this->redirect(['action' => 'index']);
+            }
+            $this->request->allowMethod(['post', 'delete']);
+            $user = $this->Users->get($id);
+            if ($this->Users->delete($user)) {
+                $this->Flash->success(__('Usuário deletado com sucesso.'));
+            } else {
+                $this->Flash->error(__('O usuário não pode ser deletado. Por favor, tente novamente.'));
+            }
+
+            return $this->redirect(['action' => 'index']);
         }
-
-        return $this->redirect(['action' => 'index']);
     }
-
 
 }
